+++ service/controller/resource/monitoring/remotewriteconfig/create.go
@@ -5,9 +5,12 @@ import (
 	"reflect"
 
 	"github.com/giantswarm/microerror"
+	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/yaml"
 
+	remotewriteconfiguration "github.com/giantswarm/prometheus-meta-operator/v2/pkg/remotewrite/configuration"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/key"
 )
 
@@ -35,7 +38,17 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 		}
 
 		if current != nil {
-			desired, err := r.desiredConfigMap(cluster, name, namespace, r.Version)
+			currentShards, err := readCurrentShardsFromConfig(*current)
+			if err != nil {
+				return microerror.Mask(err)
+			}
+
+			shards, err := r.getShardsCountForCluster(ctx, cluster, currentShards)
+			if err != nil {
+				return microerror.Mask(err)
+			}
+
+			desired, err := r.desiredConfigMap(cluster, name, namespace, r.Version, shards)
 			if err != nil {
 				return microerror.Mask(err)
 			}
@@ -52,6 +65,16 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 	r.logger.Debugf(ctx, "ensured prometheus remote write config")
 
 	return nil
+}
+
+func readCurrentShardsFromConfig(configMap corev1.ConfigMap) (int, error) {
+	remoteWriteConfig := remotewriteconfiguration.RemoteWriteConfig{}
+	err := yaml.Unmarshal([]byte(configMap.Data["values"]), &remoteWriteConfig)
+	if err != nil {
+		return 0, microerror.Mask(err)
+	}
+
+	return remoteWriteConfig.PrometheusAgentConfig.Shards, nil
 }
 
 func updateMeta(c, d metav1.Object) {